extended so system only flag equals 2 is now model metrics only

+++ Flower/flwr_balanced_benchmark.sh
@@ -33,41 +33,43 @@ if [ $SYSTEM_ONLY = "0" ]; then
   done
   echo "---------------------------------------------------------------------------------------------------------"
 fi
-echo "Benchmark system metrics"
-for (( repeat = 0; repeat < $REPEATS; repeat++ ))
-do
-  echo "Start repeat system metrics ${repeat} num clients ${NUM_CLIENTS} num rounds ${NUM_ROUNDS} and data ${DATA_NAME}"
-  rm -f timelogs/flw_logs_time.txt
-  echo "Creating server"
-  python server.py --data_path $DATA_PATH --run_repeat $repeat --num_clients $NUM_CLIENTS --num_rounds $NUM_ROUNDS --system_metrics true &
-  server_id=$!
-  # Reads in all cpu available as string
-  read cpu_available <<< $(taskset -pc $server_id | awk '{print $NF}')
-  # Converts that string into an array
-  cpu_arr=(${cpu_available//,/ })
-  # Reads in two CPU numbers for latter assignment, one for the server one for the client, because sometimes
-  # the taskset command returns a range of CPU's in form of '5-7' the cut command is needed for extracting only one number
-  cpu_num_1=$(echo ${cpu_arr[0]} | cut -d'-' -f1)
-  cpu_num_2=$(echo ${cpu_arr[1]} | cut -d'-' -f1)
-  # Bounds the CPU with 'cpu_num_1' to the server process
-  taskset -c -pa $cpu_num_1 $server_id
-  #sleep 3
-  echo "Start client"
-  python client.py --client_index 1 --data_path $DATA_PATH --run_repeat $repeat --system_metrics true &
-  client_id=$!
-  # Bind client process to cpu num 2
-  taskset -c -pa $cpu_num_2 $client_id
-  client_time_logs="timelogs/flwr_client_${DATA_NAME}_${NUM_CLIENTS}_${NUM_ROUNDS}_repeat_${repeat}.txt"
-  server_time_logs="timelogs/flwr_server_${DATA_NAME}_${NUM_CLIENTS}_${NUM_ROUNDS}_repeat_${repeat}.txt"
-  # Record memory from client and server process and log to file
-  psrecord $client_id --log $client_time_logs  --interval 0.5 &
-  psrecord $server_id --log $server_time_logs --interval 0.5 &
-  wait
-  project_name="benchmark_rounds_${NUM_ROUNDS}_${DATA_NAME}_system_metrics"
-  run_name="run_${repeat}"
-  # Read files logged from psutil to wandb
-  python ../scripts/mem_data_to_wandb.py --logs_path $client_time_logs --project_name $project_name --run_name $run_name --group_name "flwr_${NUM_CLIENTS}"  --memory_type "client"
-  python ../scripts/mem_data_to_wandb.py --logs_path $server_time_logs --project_name $project_name  --run_name $run_name --group_name "flwr_${NUM_CLIENTS}"  --memory_type "server"
-  echo "Repeat system metrics ${repeat} num clients ${NUM_CLIENTS} num rounds ${NUM_ROUNDS} and data ${DATA_NAME} complete"
-  echo "<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<"
-done
+if [ $SYSTEM_ONLY != "2" ]; then
+  echo "Benchmark system metrics"
+  for (( repeat = 0; repeat < $REPEATS; repeat++ ))
+  do
+    echo "Start repeat system metrics ${repeat} num clients ${NUM_CLIENTS} num rounds ${NUM_ROUNDS} and data ${DATA_NAME}"
+    rm -f timelogs/flw_logs_time.txt
+    echo "Creating server"
+    python server.py --data_path $DATA_PATH --run_repeat $repeat --num_clients $NUM_CLIENTS --num_rounds $NUM_ROUNDS --system_metrics true &
+    server_id=$!
+    # Reads in all cpu available as string
+    read cpu_available <<< $(taskset -pc $server_id | awk '{print $NF}')
+    # Converts that string into an array
+    cpu_arr=(${cpu_available//,/ })
+    # Reads in two CPU numbers for latter assignment, one for the server one for the client, because sometimes
+    # the taskset command returns a range of CPU's in form of '5-7' the cut command is needed for extracting only one number
+    cpu_num_1=$(echo ${cpu_arr[0]} | cut -d'-' -f1)
+    cpu_num_2=$(echo ${cpu_arr[1]} | cut -d'-' -f1)
+    # Bounds the CPU with 'cpu_num_1' to the server process
+    taskset -c -pa $cpu_num_1 $server_id
+    #sleep 3
+    echo "Start client"
+    python client.py --client_index 1 --data_path $DATA_PATH --run_repeat $repeat --system_metrics true &
+    client_id=$!
+    # Bind client process to cpu num 2
+    taskset -c -pa $cpu_num_2 $client_id
+    client_time_logs="timelogs/flwr_client_${DATA_NAME}_${NUM_CLIENTS}_${NUM_ROUNDS}_repeat_${repeat}.txt"
+    server_time_logs="timelogs/flwr_server_${DATA_NAME}_${NUM_CLIENTS}_${NUM_ROUNDS}_repeat_${repeat}.txt"
+    # Record memory from client and server process and log to file
+    psrecord $client_id --log $client_time_logs  --interval 0.5 &
+    psrecord $server_id --log $server_time_logs --interval 0.5 &
+    wait
+    project_name="benchmark_rounds_${NUM_ROUNDS}_${DATA_NAME}_system_metrics"
+    run_name="run_${repeat}"
+    # Read files logged from psutil to wandb
+    python ../scripts/mem_data_to_wandb.py --logs_path $client_time_logs --project_name $project_name --run_name $run_name --group_name "flwr_${NUM_CLIENTS}"  --memory_type "client"
+    python ../scripts/mem_data_to_wandb.py --logs_path $server_time_logs --project_name $project_name  --run_name $run_name --group_name "flwr_${NUM_CLIENTS}"  --memory_type "server"
+    echo "Repeat system metrics ${repeat} num clients ${NUM_CLIENTS} num rounds ${NUM_ROUNDS} and data ${DATA_NAME} complete"
+    echo "<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<"
+  done
+fi
